Complete code rework
